extracted import fridge tag component and logic to common
